refactor(messaging): split conflict detection modules
