+++ src/lib/messaging/applier/conflict-detection/manifest-metadata.ts
@@ -1,11 +1,12 @@
 // SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import { BUILT_IN_CHANNEL_MANIFESTS } from "../channels";
+import { BUILT_IN_CHANNEL_MANIFESTS } from "../../channels";
 
 // Map channelId to providerEnvKey values declared in built-in manifests.
-// The comparison layer uses this as the primary key set so a missing hash for
-// one required credential conservatively reports unknown-token.
+// This is the primary key set for hash comparison so a missing credential for
+// one of a channel's required credentials conservatively marks the comparison
+// as unknown-token rather than silently returning null.
 export const CHANNEL_CREDENTIAL_ENV_KEYS: Readonly<Record<string, readonly string[]>> =
   Object.fromEntries(
     BUILT_IN_CHANNEL_MANIFESTS.map((m) => [m.id, m.credentials.map((c) => c.providerEnvKey)]),
