+++ src/lib/messaging/applier/conflict-detection/plan.ts
@@ -1,13 +1,13 @@
 // SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import type { SandboxMessagingPlan } from "../manifest";
-import type { ConflictRequest } from "./conflict-detection-types";
+import type { SandboxMessagingPlan } from "../../manifest";
+import type { ConflictRequest } from "./types";
 
 /**
  * Return the channel IDs that are active (not disabled) in a compiled plan.
  * Aligns with `enabledPlanChannels()` in plan-filter.ts: a channel is active
- * only when `channel.active && !channel.disabled` and it is not in
+ * only when `channel.active && !channel.disabled` AND it is not in
  * `plan.disabledChannels`.
  */
 export function getActiveChannelIdsFromPlan(plan: SandboxMessagingPlan): string[] {
@@ -20,6 +20,11 @@ export function getActiveChannelIdsFromPlan(plan: SandboxMessagingPlan): string[
 /**
  * Return credential hashes keyed by providerEnvKey from a compiled plan,
  * optionally scoped to a single channel.
+ *
+ * Only bindings that carry a `credentialHash` are included. When `channelId`
+ * is provided only that channel's bindings are returned, which prevents
+ * hashes from other channels in the same sandbox from contaminating
+ * single-channel conflict comparisons.
  */
 export function getCredentialHashesFromPlan(
   plan: SandboxMessagingPlan,
@@ -34,12 +39,18 @@ export function getCredentialHashesFromPlan(
 }
 
 /**
- * Build conflict requests from plan credential bindings.
+ * Build a `ConflictRequest[]` from a compiled plan's credential bindings.
  *
- * Bindings are grouped by channelId, disabled channels are skipped, and
- * credentialAvailable=false bindings are omitted. A binding with no
- * credentialHash still produces a channel request so later comparison can use
- * conservative unknown-token behavior instead of dropping the channel.
+ * Groups bindings by channelId (e.g. Slack has SLACK_BOT_TOKEN and
+ * SLACK_APP_TOKEN) and excludes:
+ *   - channels in `plan.disabledChannels` (bridge is paused, not in use)
+ *   - bindings where the credential is not available (`credentialAvailable`
+ *     false) - e.g. WhatsApp, which has no host-side token provider
+ *
+ * When a binding has no `credentialHash` (e.g. a registry-only resume that
+ * did not re-run the compiler), the channel is still included with an empty
+ * `credentialHashes` map, which falls through to `"unknown-token"` conservative
+ * detection.
  */
 export function planToConflictChannelRequests(plan: SandboxMessagingPlan): ConflictRequest[] {
   const activeChannelIds = new Set(getActiveChannelIdsFromPlan(plan));
